web: add #isUsed to model drives
- The method #buildModel is extracted to a helper because it will
  be used by other hooks and/or helpers.
